+++ lib/shoulda/matchers/active_model/numericality_matchers/numeric_type_matcher.rb
@@ -2,24 +2,9 @@ module Shoulda # :nodoc:
   module Matchers
     module ActiveModel # :nodoc:
       module NumericalityMatchers
-        class OddEvenNumberMatcher # :nodoc:
-          NON_EVEN_NUMBER_VALUE = 1
-          NON_ODD_NUMBER_VALUE  = 2
-
-          def initialize(attribute, options = {})
-            @attribute = attribute
-            options[:odd]   ||= true
-            options[:even]  ||= false
-
-            if options[:odd] && !options[:even]
-              @disallow_value_matcher = DisallowValueMatcher.new(NON_ODD_NUMBER_VALUE).
-                for(@attribute).
-                with_message(:odd)
-            else
-              @disallow_value_matcher = DisallowValueMatcher.new(NON_EVEN_NUMBER_VALUE).
-                for(@attribute).
-                with_message(:even)
-            end
+        class NumericTypeMatcher
+          def initialize
+            raise NotImplementedError
           end
 
           def matches?(subject)
@@ -31,8 +16,8 @@ module Shoulda # :nodoc:
             self
           end
 
-          def allowed_types
-            'integer'
+          def allowed_type
+            raise NotImplementedError
           end
 
           def failure_message